+++ src/Controller/DoctorController.php
@@ -44,17 +44,29 @@ final class DoctorController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_doctor_show', methods: ['GET'])]
-    public function show(Doctor $doctor): Response
+    #[Route('/{id<\d+>}', name: 'app_doctor_show', methods: ['GET'])]
+    public function show(int $id, DoctorRepository $doctorRepository): Response
     {
+        $doctor = $doctorRepository->find($id);
+
+        if (!$doctor) {
+            throw $this->createNotFoundException('Doctor not found');
+        }
+
         return $this->render('doctor/show.html.twig', [
             'doctor' => $doctor,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_doctor_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Doctor $doctor, EntityManagerInterface $entityManager): Response
+    #[Route('/{id<\d+>}/edit', name: 'app_doctor_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, int $id, DoctorRepository $doctorRepository, EntityManagerInterface $entityManager): Response
     {
+        $doctor = $doctorRepository->find($id);
+
+        if (!$doctor) {
+            throw $this->createNotFoundException('Doctor not found');
+        }
+
         $form = $this->createForm(DoctorType::class, $doctor);
         $form->handleRequest($request);
 
@@ -70,9 +82,15 @@ final class DoctorController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_doctor_delete', methods: ['POST'])]
-    public function delete(Request $request, Doctor $doctor, EntityManagerInterface $entityManager): Response
+    #[Route('/{id<\d+>}', name: 'app_doctor_delete', methods: ['POST'])]
+    public function delete(Request $request, int $id, DoctorRepository $doctorRepository, EntityManagerInterface $entityManager): Response
     {
+        $doctor = $doctorRepository->find($id);
+
+        if (!$doctor) {
+            throw $this->createNotFoundException('Doctor not found');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$doctor->getId(), $request->request->get('_token'))) {
             $entityManager->remove($doctor);
             $entityManager->flush();
@@ -81,31 +99,32 @@ final class DoctorController extends AbstractController
         return $this->redirectToRoute('app_doctor_index', [], Response::HTTP_SEE_OTHER);
     }
 
-    #[Route('/home/{id}', name: 'app_doctor_home', methods: ['GET'])]
+    #[Route('/home/{id<\d+>}', name: 'app_doctor_home', methods: ['GET'])]
     public function home(DoctorRepository $doctorRepository, int $id, RequestStack $requestStack): Response
     {
         $doctor = $doctorRepository->find($id);
-    
+
         if (!$doctor) {
             throw $this->createNotFoundException('Doctor not found');
         }
 
         $session = $requestStack->getSession();
         $session->set('doctorId', $id);
-    
+
         return $this->render('doctor/home.html.twig', [
-            'doctor_name' => $doctor->getName(),
+            'doctor_name' => $doctor->getNom(),
             'template' => 'template1',
         ]);
     }
+
     #[Route('/users', name: 'app_doctor_dashboard')]
     public function userlist(): Response
     {
         $user = $this->getUser();
         return $this->render('doctor/dashboard.html.twig', [
             'user' => $user,
+            'template' => 'template1',
+
         ]);
     }
-    
-
 }
